+++ src/main/java/com/example/CargoTracking/model/DomesticShipment.java
@@ -76,7 +76,7 @@ public class DomesticShipment {
     private boolean escalationFlagThree;
     private LocalDate createdAt;
     private LocalDate updatedAt;
-    private Double damage;
+    private Integer damage;
     private String damageAwbs;
     @ManyToOne
     private User createdBy;
